requirements.txt
claude-agent-sdk
openai
numpy
scipy
pandas
matplotlib
seaborn
requests

# Model fitting and statistics
lmfit
statsmodels
scikit-learn
xgboost
prophet

# Disease transmission modeling
laser-generic

# Bayesian modeling (optional, heavier dependency)
# pymc
